+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- New: Add option to upload and download the database file (SQLite only).
 - Fix: Update column number for the ICAO code for Our Airports database source.
 - Fix: Recalculate night time even the field is not empty.
 - New: Added an option to hide time fields on the Flight Record page.